+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 ### Fixed
 - Menu: Menu title now updates correctly on language change.
 - Issue #1411: Menu: Fixed duplicated breadcrumbs in Subject Layer selection after closing LayerInfo.
+- GraphicalSelect: Fix alert message popping up if not all values given in geographicValues are used in the options props
 
 ---